ci: splits frontend e2e tests into 3 different groups
