+++ .github/workflows/task_fe_unit_tests.yml
@@ -5,7 +5,7 @@ on:
 
 jobs:
   unit:
-    name: 'Frontend unit tests'
+    name: 'vitest'
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
