+++ frontend/app/tests/e2e/specs/app/assets.spec.ts
