+++ frontend/app/tests/e2e/specs/app/defi.spec.ts
