+++ frontend/app/tests/e2e/specs/app/settings.spec.ts
